chore: extract filter utils to filter file from common/index
